data_acquisition/requirements.txt
fastapi
uvicorn
redis
requests
httpx
python-dotenv
yfinance
pydantic
prometheus_client
rq 